# src/routes/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Darve Server API",
    "description": "API documentation for the Darve Server including users, posts, notifications, auth, discussions, tasks, wallet, and follows",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "",
      "description": "Development server"
    }
  ],
  "security": [],
  "paths": {
...
    "/api/tasks/received": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Get received task requests",
        "description": "Get task requests received by the current user",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by participant status",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["PENDING", "ACCEPTED", "REJECTED", "COMPLETED"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of received task requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskRequestView"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/tasks/given": {
      "get": {
        "tags": ["Tasks"],
        "summary": "Get given task requests",
        "description": "Get task requests created by the current user",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of given task requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskRequestView"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/tasks/{task_id}/accept": {
      "post": {
        "tags": ["Tasks"],
        "summary": "Accept task request",
        "description": "Accept a task request",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "description": "ID of the task to accept",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task accepted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    },
    "/api/tasks/{task_id}/reject": {
      "post": {
        "tags": ["Tasks"],
        "summary": "Reject task request",
        "description": "Reject a task request",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "description": "ID of the task to reject",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task rejected successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    },
...
    "/api/tasks/{task_id}/donor": {
      "post": {
        "tags": ["Tasks"],
        "summary": "Add or update donor",
        "description": "Add or update a donor for a task",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "description": "ID of the task",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRequestOfferInput"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/TaskRequestOfferInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Donor added/updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TaskRequestOfferInput": {
        "type": "object",
        "required": ["amount"],
        "properties": {
          "amount": {
            "type": "integer",
            "format": "int64",
            "minimum": 100,
            "description": "Offer amount"
          }
        }
      },
...
      "TaskRequestView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Task request ID"
          },
          "created_by": {
            "$ref": "#/components/schemas/UserView"
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskRequestViewParticipant"
            },
            "description": "Task participants"
          },
          "request_txt": {
            "type": "string",
            "description": "Task request description"
          },
          "donors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskRequestDonorView"
            },
            "description": "Task donors"
          },
          "reward_type": {
            "$ref": "#/components/schemas/RewardType"
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencySymbol"
          },
          "wallet_id": {
            "type": "string",
            "description": "Wallet ID"
          }
        }
      },
      "TaskRequestViewParticipant": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserView"
          },
          "status": {
            "$ref": "#/components/schemas/TaskParticipantStatus"
          }
        }
      },
      "TaskRequestDonorView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Donor ID"
          },
          "user": {
            "$ref": "#/components/schemas/UserView"
          },
          "amount": {
            "type": "integer",
            "format": "int64",
            "description": "Donation amount"
          }
        }
      },
      "UserView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID"
          },
          "username": {
            "type": "string",
            "description": "Username"
          },
          "full_name": {
            "type": "string",
            "description": "Full name"
          }
        }
      },
      "TaskParticipantStatus": {
        "type": "string",
        "enum": ["Requested", "Rejected", "Accepted", "Delivered"],
        "description": "Status of task participant"
      },
      "RewardType": {
        "type": "string",
        "description": "Type of reward for the task"
      },
      "CurrencySymbol": {
        "type": "string",
        "description": "Currency symbol"
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt"
      }
    }
  },
...
}